moved cnt core logic to phichain-game
